+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Documented custom userdata types (Job and Timer Objects)
   - Coming with fix is also adding the return types for some functions that were missing it
 
+### Fixed
+- `hilbish.which` not working correctly with aliases
+
 ## [2.0.1] - 2022-12-28
 ### Fixed
 - Corrected documentation for hooks, removing outdated `command.no-perm`